Add files via upload
Customer churn rate is a metric that measures the percentage of customers who end their relationship with a company within a particular period [1].

The research paper’s goal is to create and develop an automatic churn prediction system by using big data technology. This particular work is focused on the telecommunications industry; however, this study might be replicated in other similar situations, such as employee turnover, where the primary goal is to try to retain employees.