Add files via upload
Graphs are general and powerful data structures that can be used to represent diverse kinds of objects [1]. Thanks to their easy generalisation, they have been progressively crucial in modelling complex structures in a broad variety of applications. In fact, the majority of the data in the world at the moment can be classified as unstructured and graph have become a very practical tool to represent this kind of data. Examples can be found in semi-structured data such as XML and HTML, chemical compounds, biological sequences, computer vision, Web analysis and Social Networks.